Moved full screen quad code to its own file
