+++ src/core/mod.rs
@@ -1,6 +1,7 @@
 pub mod buffer;
 pub mod program;
 pub mod rendertarget;
+pub mod full_screen_quad;
 mod shader;
 pub mod state;
 pub mod texture;
